New package "otel"
Provides utilities to instrument applications using OpenTelemetry.

Proper instrumentation is crucial to monitor system performance,
behavior, and to detect problems, regressions and bugs. This practice
is usually referred to as observability. The 3 pillars of observability
are: logs, metrics and traces. All these details must be generated,
collected and analyzed in a logical and consistent manner.

This package simplify proper instrumentation of systems by integrating
all 3 data sources into a single interface with the following
characteristics: 1) easy to setup; 2) easy to use; 3) consistent
(and mostly automatic) behavior.