New package "crypto/shamir"
Provides a simple implementation of Shamir's Secret Sharing algorithm.

Shamir's Secret Sharing is a cryptographic algorithm created by Adi
Shamir. It is a form of secret sharing, where a secret is divided into
several unique parts (shares). To reconstruct the original secret, a
minimum number (threshold) of parts is required. In the threshold scheme
this number is less than the total number of parts. Otherwise all
participants are needed to reconstruct the original secret.