New package "crypto/pow"
Provides an implementation of the Proof-of-Work algorithm.

The current Proof-of-Work algorithms are based on the original HashCash
proposal. HashCash was originally designed as a mechanism to throttle
systematic abuse of un-metered internet resources such as email, and
anonymous re-mailers in May 1997. The algorithm is based on utilizing a
cost-function that must be efficiently verifiable, but parameterizable
expensive to compute.