Sanitize log input
If unsanitized user input is written to a log entry, a malicious user
may be able to forge new log entries. Forgery can occur if a user
provides some input with characters that are interpreted when the log
output is displayed. If the log is displayed as a plain text file, then
new line characters can be used by a malicious user. If the log is
displayed as HTML, then arbitrary HTML may be included to spoof log
entries.

More information:
https://owasp.org/www-community/attacks/Log_Injection