Error Reporting
Add support for a new "error reporting" API on "otel" instrumentation
package. This interface can be used to submit exception details to 3rd
party service for processing and tracking.

In addition to the new APIs the package has been refactored to provide
a cleaner and more consistent set of primitives and APIs.

The update includes a reference error reporting implementation for
Sentry.

More information:
https://sentry.io/welcome/