New package "net/sse"
The package provides a "Server-Sent Events" implementation.

Traditionally, a web page has to send a request to the server to receive
new data; that is, the page requests data from the server. With SSE it's
possible for a server to send new data to a web page at any time, by
pushing messages to the web page.

More information:
https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events