DID Well Known Configuration
Making it possible to connect existing systems and Decentralized
Identifiers (DIDs) is an important undertaking that can aid in
bootstrapping adoption and usefulness of DIDs. One such form of
connection is the ability of a DID controller to prove they are
the same entity that controls an Internet domain.

The DID Configuration resource provides proof of a bi-directional
relationship between the controller of an Internet domain and a DID
via cryptographically verifiable signatures that are linked to a
DID's key material.

The DID Configuration resource MUST exist at the domain root, in
the IETF 8615 Well-Known Resource directory
`/.well-known/did-configuration`.

More information:
https://identity.foundation/specs/did-configuration/