OTEL: New GORM plugin
The package provides an OpenTelemetry instrumentation for the GORM
library. The new implementation provides additional features such as
defining specific errors to be ignored when reporting span status. This
is useful to avoid unnecessary noise in the tracing system, for example
when a record is not found or a request is cancelled by the user.

More information:
https://github.com/go-gorm/opentelemetry