feat: JWK Thumbprint
Add support for RFC 7638 to the `jwk` package. The specification defines
a method for computing a hash value over a JSON Web Key (JWK). It
defines which fields in a JWK are used in the hash computation, the
method of creating a canonical form for those fields, and how to convert
the resulting Unicode string into a byte sequence to be hashed. The
resulting hash value can be used for identifying or selecting the key
represented by the JWK that is the subject of the thumbprint.

More information:
https://www.rfc-editor.org/rfc/rfc7638.html