feat: New CSRF middleware
In a cross-site request forgery (CSRF) attack, an attacker tricks the user
or the browser into making an HTTP request to the target site from a malicious
site. The request includes the user's credentials and causes the server to
carry out some harmful action, thinking that the user intended it.

More information:
https://developer.mozilla.org/en-US/docs/Web/Security/Attacks/CSRF